% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PbPbRedux.R
\name{process}
\alias{process}
\title{Pb-Pb data processing}
\usage{
process(samples, blanks, spikes, cblanks)
}
\arguments{
\item{samples}{data frame with sample data. For an example, see
\code{system.file("samples1.csv",package="PbPbRedux")}}

\item{blanks}{data frame with blank data. For an example, see
\code{system.file("blanks1.csv",package="PbPbRedux")}}

\item{spikes}{data frame with spike data. For an example, see
\code{system.file("spikes.csv",package="PbPbRedux")}}

\item{cblanks}{data frame with replicate blank data. For an
example, see
\code{system.file("blanks2.csv",package="PbPbRedux")}}
}
\value{
a table with Pb concentrations, Pb/Pb ratios and error
correlations
}
\description{
High level functionthat takes samples, blanks and
spikes as input and produces a table of Pb/Pb ratios,
uncertainties and error correlations as output
}
\examples{
library(PbPbRedux)

s1 <- system.file("samples1.csv",package="PbPbRedux")
s2 <- system.file("samples2.csv",package="PbPbRedux")
b1 <- system.file("blanks1.csv",package="PbPbRedux")
b2 <- system.file("blanks2.csv",package="PbPbRedux")
sp <- system.file("spikes.csv",package="PbPbRedux")

spikes <- read.csv(sp,header=TRUE)

# example 1: all samples use the same blank:
samples <- read.csv(s1,header=TRUE)
blanks <- read.csv(b1,header=TRUE)
tab <- process(samples,blanks,spikes)

# example 2: each aliquot has its own blank:
samples <- read.csv('s2.csv',header=TRUE)
blanks <- read.csv('b2.csv',header=TRUE)
tab <- process(samples,blanks,spikes)

# example 3: individual blanks with shared covariance matrix:
samples <- read.csv('s2.csv',header=TRUE)
cblanks <- read.csv('b1.csv',header=TRUE)
blanks <- read.csv('b2.csv',header=TRUE)
tab <- process(samples,blanks,spikes,cblanks)
}
